Fixes from DBIC field-test site and code rodeo coding session (AKA reaching convergence)

+++ heudiconv/utils.py
@@ -1,5 +1,5 @@
 """Utility objects and functions"""
-
+import hashlib
 import os
 import tempfile
 import json
@@ -183,6 +183,12 @@ def load_json(filename):
     return data
 
 
+def assure_no_file_exists(path):
+    """Check if file or symlink (git-annex?) exists, and if so -- remove"""
+    if os.path.lexists(path):
+        os.unlink(path)
+
+
 def save_json(filename, data, indent=4):
     """Save data to a json file
 
@@ -194,6 +200,7 @@ def save_json(filename, data, indent=4):
         Dictionary to save in json file.
 
     """
+    assure_no_file_exists(filename)
     with open(filename, 'w') as fp:
         fp.write(_canonical_dumps(data, sort_keys=True, indent=indent))
 
@@ -267,12 +274,17 @@ def load_heuristic(heuristic_file):
     return mod
 
 
-def safe_copyfile(src, dest):
+def safe_copyfile(src, dest, overwrite=False):
     """Copy file but blow if destination name already exists
     """
     if op.isdir(dest):
         dest = op.join(dest, op.basename(src))
     if op.lexists(dest):
+        if not overwrite:
+            raise RuntimeError(
+                "was asked to copy %s but destination already exists: %s"
+                % (src, dest)
+            )
         os.unlink(dest)
     shutil.copyfile(src, dest)
 
@@ -333,3 +345,8 @@ def clear_temp_dicoms(item_dicoms):
         and op.exists(str(tmp))):
         # clean up directory holding dicoms
         shutil.rmtree(str(tmp))
+
+
+def file_md5sum(filename):
+    with open(filename, 'rb') as f:
+        return hashlib.md5(f.read()).hexdigest()
